refactor(logo) - rename logo component
