+++ src/components/layout/AppHeader/AppHeader.tsx
@@ -1,13 +1,13 @@
 import classNames from 'classnames'
 
-import { Logo, Sitemap } from '@/components/shared'
+import { WebsiteLogo, Sitemap } from '@/components/shared'
 
 import styles from './AppHeader.module.scss'
 
 const AppHeader = () => {
   return (
     <header className={classNames(styles.appHeader, 'fullRow')}>
-      <Logo />
+      <WebsiteLogo />
 
       <nav>
         <Sitemap />
